Modularized code by seperating Helper Functions
