+++ Assignment 11/IPC/MessageQueue.c
@@ -4,6 +4,7 @@
 #include <sys/ipc.h>
 #include <sys/msg.h>
 #include <sys/wait.h>
+#include "IPC_HELPER.h"
 
 #define MESSAGE_KEY 1234
 #define SIZE 50
@@ -13,66 +14,6 @@ typedef struct Message {
     int size;
     int array[SIZE];
 } Message;
-
-int* readArray(int *size) {
-    printf("Enter the size of array: ");
-    scanf("%d", size);
-
-    int* array = malloc((*size) * sizeof(int));
-    if (!array) 
-    {
-        printf("Memory Allocation Failed!\n");
-        exit(1);
-    }
-
-    printf("Enter elements of array: ");
-    for (int index = 0; index < *size; index++)
-    {
-        scanf("%d", &array[index]);
-    }
-
-    return array;
-}
-
-void display(int* array, int size) {
-    for (int index = 0; index < size; index++)
-    {
-        printf("%d ", array[index]);
-    }
-    printf("\n");
-}
-
-void swap(int *a, int *b) {
-    int temp = *a;
-    *a = *b;
-    *b = temp;
-}
-
-int partition(int* array, int low, int high) {
-    int pivot = array[high];
-    int index = low - 1;
-
-    for (int currentIndex = low; currentIndex < high; currentIndex++)
-    {
-        if (array[currentIndex] < pivot)
-        {
-            index++;
-            swap(&array[currentIndex], &array[index]);
-        }
-    }
-
-    swap(&array[index + 1], &array[high]);
-    return index + 1;
-}
-
-void quickSort(int *array, int low, int high) {
-    if (low < high)
-    {
-        int partitionIndex = partition(array, low, high);
-        quickSort(array, low, partitionIndex - 1);
-        quickSort(array, partitionIndex + 1, high);
-    }
-}
 
 void sendMessage(int messageId, long messageType, int *array, int size) {
     Message message;
@@ -112,7 +53,7 @@ void parentProcess(int messageId, int* array, int* size) {
     receiveMessage(messageId, 2, array, size);
 
     printf("Array After Sorting: \n");
-    display(array, *size);
+    displayArray(array, *size);
 
     msgctl(messageId, IPC_RMID, NULL);
 }
@@ -129,7 +70,7 @@ int main() {
     }
     
     printf("Array Before Sorting: \n");
-    display(array, size);
+    displayArray(array, size);
 
     pid_t pid = fork();
