+++ Assignment 11/IPC/PipeIPC.c
@@ -2,66 +2,7 @@
 #include <stdlib.h>
 #include <unistd.h>
 #include <sys/wait.h>
-
-int* readArray(int *size) {
-    printf("Enter the size of array: ");
-    scanf("%d", size);
-
-    int* array = malloc((*size) * sizeof(int));
-    if (!array) 
-    {
-        printf("Memory Allocation Failed!\n");
-        exit(1);
-    }
-
-    printf("Enter elements of array: ");
-    for (int index = 0; index < *size; index++)
-    {
-        scanf("%d", &array[index]);
-    }
-
-    return array;
-}
-
-void display(int* array, int size) {
-    for (int index = 0; index < size; index++)
-    {
-        printf("%d ", array[index]);
-    }
-    printf("\n");
-}
-
-void swap(int *a, int *b) {
-    int temp = *a;
-    *a = *b;
-    *b = temp;
-}
-
-int partition(int* array, int low, int high) {
-    int pivot = array[high];
-    int index = low - 1;
-
-    for (int currentIndex = low; currentIndex < high; currentIndex++)
-    {
-        if (array[currentIndex] < pivot)
-        {
-            index++;
-            swap(&array[currentIndex], &array[index]);
-        }
-    }
-
-    swap(&array[index + 1], &array[high]);
-    return index + 1;
-}
-
-void quickSort(int *array, int low, int high) {
-    if (low < high)
-    {
-        int partitionIndex = partition(array, low, high);
-        quickSort(array, low, partitionIndex - 1);
-        quickSort(array, partitionIndex + 1, high);
-    }
-}
+#include "IPC_HELPER.h"
 
 void sendData(int fileDescriptor, int *array, int size) {
     if (write(fileDescriptor, &size, sizeof(int)) != sizeof(int))
@@ -116,7 +57,7 @@ void parentProcess(int *array, int size, int parent2child[2], int child2parent[2
     receiveData(child2parent[0], array, &size);
 
     printf("Array After Sorting: \n");
-    display(array, size);
+    displayArray(array, size);
 
     close(parent2child[1]);
     close(child2parent[0]);
@@ -127,7 +68,7 @@ int main() {
     int* array = readArray(&size);
     
     printf("Array Before Sorting: \n");
-    display(array, size);
+    displayArray(array, size);
 
     int parent2child[2], child2parent[2];
     if (pipe(parent2child) == -1 || pipe(child2parent) == -1)
